feat(ACI): Refactor ACI module (#505)
* feat(Selector): If default account `address` not provided use the first

* feat(ACI): Refactor ACI. Split to separated files. Fix test.

* feat(ACI): Refactor building of contract instance method, add ability to init instance without client

* feat(ACI): Handle ACI without init function

* chore(ACI): Cleanup code

* feat(ACI): Refactor ACI module. Fix bug with empty init function. Automatically decide to use on-chai

* fix(ACI): Fix options composition when building function

* chore(docs): Fix ACI docs example
