+++ es/ae/universal.js
@@ -26,9 +26,11 @@ import Ae from './'
 import Account from '../account/memory'
 import Chain from '../chain/node'
 import Aens from './aens'
-import Contract from './contract'
 import Transaction from '../tx/tx'
 import Oracle from './oracle'
+import Selector from '../account/selector'
+import Accounts from '../accounts'
+import Contract from './contract'
 
 /**
  * Universal Stamp
@@ -41,7 +43,12 @@ import Oracle from './oracle'
  * @param {Object} [options={}] - Initializer object
  * @return {Object} Universal instance
  */
-const Universal = Ae.compose(Account, Chain, Transaction, Aens, Contract, Oracle, {
+export const Universal = Ae.compose(Account, Chain, Transaction, Aens, Contract, Oracle, {
+  init () {},
+  props: { process: {} }
+})
+
+export const UniversalWithAccounts = Ae.compose(Accounts, Chain, Transaction, Aens, Contract, Oracle, Selector, {
   init () {},
   props: { process: {} }
 })
