--- es/contract/aci.js
@@ -1,540 +1,0 @@
-/* eslint-disable no-unused-vars */
-/*
- * ISC License (ISC)
- * Copyright (c) 2018 aeternity developers
- *
- *  Permission to use, copy, modify, and/or distribute this software for any
- *  purpose with or without fee is hereby granted, provided that the above
- *  copyright notice and this permission notice appear in all copies.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
- *  REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
- *  AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
- *  INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
- *  LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
- *  OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
- *  PERFORMANCE OF THIS SOFTWARE.
- */
-
-/**
- * ContractACI module
- *
- * @module @aeternity/aepp-sdk/es/contract/aci
- * @export ContractACI
- * @example import ContractACI from '@aeternity/aepp-sdk/es/contract/aci'
- */
-import Joi from 'joi-browser'
-
-import AsyncInit from '../utils/async-init'
-import * as R from 'ramda'
-
-const SOPHIA_TYPES = [
-  'int',
-  'string',
-  'tuple',
-  'address',
-  'bool',
-  'list',
-  'map',
-  'record',
-  'option',
-  'oracle',
-  'oracleQuery',
-  'hash',
-  'signature',
-  'bytes'
-].reduce((acc, type) => ({ ...acc, [type]: type }), {})
-
-/**
- * Transform decoded data to JS type
- * @param aci
- * @param result
- * @param transformDecodedData
- * @return {*}
- */
-function transformDecodedData (aci, result, { skipTransformDecoded = false, addressPrefix = 'ak', bindings } = {}) {
-  if (skipTransformDecoded) return result
-  const { t, generic } = readType(aci, { bindings })
-
-  switch (t) {
-    case SOPHIA_TYPES.bool:
-      return !!result
-    case SOPHIA_TYPES.address:
-      return result === 0
-        ? 0
-        : result
-    case SOPHIA_TYPES.hash:
-    case SOPHIA_TYPES.bytes:
-    case SOPHIA_TYPES.signature:
-      return result.split('#')[1]
-    case SOPHIA_TYPES.map:
-      const [keyT, valueT] = generic
-      return result
-        .reduce(
-          (acc, [key, val]) => {
-            key = transformDecodedData(keyT, key, { bindings })
-            val = transformDecodedData(valueT, val, { bindings })
-            acc.push([key, val])
-            return acc
-          },
-          []
-        )
-    case SOPHIA_TYPES.option:
-      if (result === 'None') return undefined
-      const [[variantType, [value]]] = Object.entries(result)
-      return variantType === 'Some' ? transformDecodedData(generic, value, { bindings }) : undefined
-    case SOPHIA_TYPES.list:
-      return result.map((value) => transformDecodedData(generic, value, { bindings }))
-    case SOPHIA_TYPES.tuple:
-      return result.map((value, i) => { return transformDecodedData(generic[i], value, { bindings }) })
-    case SOPHIA_TYPES.record:
-      const genericMap = generic.reduce((acc, val) => ({ ...acc, [val.name]: { type: val.type } }), {})
-      return Object.entries(result).reduce(
-        (acc, [name, value]) =>
-          ({
-            ...acc,
-            [name]: transformDecodedData(genericMap[name].type, value, { bindings })
-          }),
-        {}
-      )
-  }
-  return result
-}
-
-/**
- * Transform JS type to Sophia-type
- * @param type
- * @param value
- * @param compilerVersion
- * @param bindings
- * @return {string}
- */
-async function transform (type, value, { compilerVersion, bindings } = {}) {
-  let { t, generic } = readType(type, { bindings })
-
-  // contract TestContract = ...
-  // fn(ct: TestContract)
-  if (typeof value === 'string' && value.slice(0, 2) === 'ct') t = SOPHIA_TYPES.address // Handle Contract address transformation
-
-  switch (t) {
-    case SOPHIA_TYPES.string:
-      return `"${value}"`
-    case SOPHIA_TYPES.list:
-      return `[${await Promise.all(value.map(async el => transform(generic, el, { compilerVersion, bindings })))}]`
-    case SOPHIA_TYPES.tuple:
-      return `(${await Promise.all(value.map(async (el, i) => transform(generic[i], el, {
-        compilerVersion,
-        bindings
-      })))})`
-    case SOPHIA_TYPES.option:
-      const optionV = await value.catch(e => undefined)
-      return optionV === undefined ? 'None' : `Some(${await transform(generic, optionV, {
-        compilerVersion,
-        bindings
-      })})`
-    case SOPHIA_TYPES.hash:
-    case SOPHIA_TYPES.bytes:
-    case SOPHIA_TYPES.signature:
-      return `#${typeof value === 'string' ? value : Buffer.from(value).toString('hex')}`
-    case SOPHIA_TYPES.record:
-      return `{${await generic.reduce(
-        async (acc, { name, type }, i) => {
-          acc = await acc
-          acc += `${i !== 0 ? ',' : ''}${name} = ${await transform(type, value[name], {
-            compilerVersion,
-            bindings
-          })}`
-          return acc
-        },
-        ''
-      )}}`
-    case SOPHIA_TYPES.map:
-      return transformMap(value, generic, { compilerVersion, bindings })
-  }
-
-  return `${value}`
-}
-
-async function transformMap (value, generic, { compilerVersion, bindings }) {
-  if (value instanceof Map) {
-    value = Array.from(value.entries())
-  }
-  if (!Array.isArray(value) && value instanceof Object) {
-    value = Object.entries(value)
-  }
-
-  return `{${await value
-    .reduce(
-      async (acc, [key, value], i) => {
-        acc = await acc
-        if (i !== 0) acc += ','
-        acc += `[${await transform(generic[0], key, {
-          compilerVersion,
-          bindings
-        })}] = ${await transform(generic[1], value, { compilerVersion, bindings })}`
-        return acc
-      },
-      ``
-    )
-  }}`
-}
-
-function linkTypeDefs (t, bindings) {
-  const [_, typeDef] = t.split('.')
-  const aciType = [
-    ...bindings.typedef,
-    { name: 'state', typedef: bindings.state }
-  ].find(({ name }) => name === typeDef)
-  return aciType.typedef
-}
-/**
- * Parse sophia type
- * @param type
- * @param returnType
- * @return {*}
- */
-function readType (type, { bindings } = {}) {
-  let [t] = Array.isArray(type) ? type : [type]
-
-  // Link State and typeDef
-  if (typeof t === 'string' && t.indexOf(bindings.contractName) !== -1) {
-    t = linkTypeDefs(t, bindings)
-  }
-  // Map, Tuple, List, Record, Bytes
-  if (typeof t === 'object') {
-    const [[baseType, generic]] = Object.entries(t)
-    return { t: baseType, generic }
-  }
-
-  // Base types
-  if (typeof t === 'string') return { t }
-}
-const customJoi = Joi.extend((joi) => ({
-  name: 'binary',
-  base: joi.any(),
-  pre (value, state, options) {
-    if (options.convert && typeof value === 'string') {
-      try {
-        return Buffer.from(value, 'hex')
-      } catch (e) { return undefined }
-    }
-
-    return Buffer.from(value)
-  },
-  rules: [
-    {
-      name: 'bufferCheck',
-      params: {
-        size: joi.number().required()
-      },
-      validate (params, value, state, options) {
-        value = value === 'string' ? Buffer.from(value, 'hex') : Buffer.from(value)
-        if (!Buffer.isBuffer(value)) {
-          return this.createError('binary.base', { value }, state, options)
-        }
-        if (value.length !== params.size) {
-          return this.createError('binary.bufferCheck', { value, size: params.size }, state, options)
-        }
-
-        return value
-      }
-    }
-  ]
-}))
-/**
- * Prepare Joi validation schema for sophia types
- * @param type
- * @param bindings
- * @return {Object} JoiSchema
- */
-function prepareSchema (type, { bindings } = {}) {
-  let { t, generic } = readType(type, { bindings })
-  if (!Object.keys(SOPHIA_TYPES).includes(t)) t = SOPHIA_TYPES.address // Handle Contract address transformation
-  switch (t) {
-    case SOPHIA_TYPES.int:
-      return Joi.number().error(getJoiErrorMsg)
-    case SOPHIA_TYPES.string:
-      return Joi.string().error(getJoiErrorMsg)
-    case SOPHIA_TYPES.address:
-      return Joi.string().regex(/^(ak_|ct_|ok_|oq_)/).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.bool:
-      return Joi.boolean().error(getJoiErrorMsg)
-    case SOPHIA_TYPES.list:
-      return Joi.array().items(prepareSchema(generic, { bindings })).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.tuple:
-      return Joi.array().ordered(generic.map(type => prepareSchema(type, { bindings }).required())).label('Tuple argument').error(getJoiErrorMsg)
-    case SOPHIA_TYPES.record:
-      return Joi.object(
-        generic.reduce((acc, { name, type }) => ({ ...acc, [name]: prepareSchema(type, { bindings }) }), {})
-      ).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.hash:
-      return customJoi.binary().bufferCheck(32).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.bytes:
-      return customJoi.binary().bufferCheck(generic).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.signature:
-      return customJoi.binary().bufferCheck(64).error(getJoiErrorMsg)
-    case SOPHIA_TYPES.option:
-      return Joi.object().type(Promise).error(getJoiErrorMsg)
-    // @Todo Need to transform Map to Array of arrays before validating it
-    // case SOPHIA_TYPES.map:
-    //   return Joi.array().items(Joi.array().ordered(generic.map(type => prepareSchema(type))))
-    default:
-      return Joi.any()
-  }
-}
-
-function getJoiErrorMsg (errors) {
-  return errors.map(err => {
-    const { path, type, context } = err
-    let value = context.hasOwnProperty('value') ? context.value : context.label
-    value = typeof value === 'object' ? JSON.stringify(value).slice(1).slice(0, -1) : value
-    switch (type) {
-      case 'string.base':
-        return ({ ...err, message: `Value "${value}" at path: [${path}] not a string` })
-      case 'number.base':
-        return ({ ...err, message: `Value "${value}" at path: [${path}] not a number` })
-      case 'boolean.base':
-        return ({ ...err, message: `Value "${value}" at path: [${path}] not a boolean` })
-      case 'array.base':
-        return ({ ...err, message: `Value "${value}" at path: [${path}] not a array` })
-      case 'object.base':
-        return ({ ...err, message: `Value '${value}' at path: [${path}] not a object` })
-      case 'object.type':
-        return ({ ...err, message: `Value '${value}' at path: [${path}] not a ${context.type}` })
-      case 'binary.bufferCheck':
-        return ({ ...err, message: `Value '${Buffer.from(value).toString('hex')}' at path: [${path}] not a ${context.size} bytes` })
-      default:
-        return err
-    }
-  })
-}
-
-function validateArguments (aci, params) {
-  const validationSchema = Joi.array().ordered(
-    aci.arguments
-      .map(({ type }, i) => prepareSchema(type, { bindings: aci.bindings }).label(`[${params[i]}]`))
-  ).label('Argument')
-  const { error } = Joi.validate(params, validationSchema, { abortEarly: false })
-  if (error) {
-    throw error
-  }
-}
-
-/**
- * Validated contract call arguments using contract ACI
- * @function validateCallParams
- * @rtype (aci: Object, params: Array) => Object
- * @param {Object} aci Contract ACI
- * @param {Array} params Contract call arguments
- * @param compilerVersion
- * @return Promise{Array} Object with validation errors
- */
-async function prepareArgsForEncode (aci, params, { compilerVersion } = {}) {
-  if (!aci) return params
-  // Validation
-  validateArguments(aci, params)
-  const bindings = aci.bindings
-  // Cast argument from JS to Sophia type
-  return Promise.all(aci.arguments.map(async ({ type }, i) => transform(type, params[i], {
-    compilerVersion,
-    bindings
-  })))
-}
-
-/**
- * Get function schema from contract ACI object
- * @param {Object} aci Contract ACI object
- * @param {String} name Function name
- * @return {Object} function ACI
- */
-function getFunctionACI (aci, name) {
-  const fn = aci.functions.find(f => f.name === name)
-  if (!fn && name !== 'init') throw new Error(`Function ${name} doesn't exist in contract`)
-
-  return {
-    ...fn,
-    bindings: {
-      state: aci.state,
-      typedef: aci.type_defs,
-      contractName: aci.name
-    }
-  }
-}
-
-/**
- * Generate contract ACI object with predefined js methods for contract usage
- * @alias module:@aeternity/aepp-sdk/es/contract/aci
- * @param {String} source Contract source code
- * @param {Object} [options] Options object
- * @param {Object} [options.aci] Contract ACI
- * @param {Object} [options.contractAddress] Contract address
- * @param {Object} [options.opt] Contract options
- * @return {ContractInstance} JS Contract API
- * @example
- * const contractIns = await client.getContractInstance(sourceCode)
- * await contractIns.compile()
- * await contractIns.deploy([321])
- * const callResult = await contractIns.call('setState', [123])
- * const staticCallResult = await contractIns.call('setState', [123], { callStatic: true })
- */
-async function getContractInstance (source, { aci, contractAddress, opt } = {}) {
-  aci = aci || await this.contractGetACI(source)
-  const defaultOptions = {
-    skipArgsConvert: false,
-    skipTransformDecoded: false,
-    callStatic: false,
-    deposit: 0,
-    gasPrice: 1000000000, // min gasPrice 1e9
-    amount: 0,
-    gas: 1600000 - 21000,
-    top: null, // using for contract call static
-    waitMined: true,
-    verify: false
-  }
-  const instance = {
-    interface: aci.interface,
-    aci: aci.encoded_aci.contract,
-    source,
-    compiled: null,
-    deployInfo: { address: contractAddress },
-    options: R.merge(defaultOptions, opt),
-    compilerVersion: this.compilerVersion,
-    setOptions (opt) {
-      this.options = R.merge(this.options, opt)
-    }
-  }
-  /**
-   * Compile contract
-   * @alias module:@aeternity/aepp-sdk/es/contract/aci
-   * @rtype () => ContractInstance: Object
-   * @return {ContractInstance} Contract ACI object with predefined js methods for contract usage
-   */
-  instance.compile = compile(this).bind(instance)
-  /**
-   * Deploy contract
-   * @alias module:@aeternity/aepp-sdk/es/contract/aci
-   * @rtype (init: Array, options: Object = { skipArgsConvert: false }) => ContractInstance: Object
-   * @param {Array} init Contract init function arguments array
-   * @param {Object} [options={}] options Options object
-   * @param {Boolean} [options.skipArgsConvert=false] Skip Validation and Transforming arguments before prepare call-data
-   * @return {ContractInstance} Contract ACI object with predefined js methods for contract usage
-   */
-  instance.deploy = deploy(this).bind(instance)
-  /**
-   * Call contract function
-   * @alias module:@aeternity/aepp-sdk/es/contract/aci
-   * @rtype (init: Array, options: Object = { skipArgsConvert: false, skipTransformDecoded: false, callStatic: false }) => CallResult: Object
-   * @param {String} fn Function name
-   * @param {Array} params Array of function arguments
-   * @param {Object} [options={}] Array of function arguments
-   * @param {Boolean} [options.skipArgsConvert=false] Skip Validation and Transforming arguments before prepare call-data
-   * @param {Boolean} [options.skipTransformDecoded=false] Skip Transform decoded data to JS type
-   * @param {Boolean} [options.callStatic=false] Static function call
-   * @return {Object} CallResult
-   */
-  instance.call = call(this).bind(instance)
-
-  instance.methods = instance
-    .aci
-    .functions
-    .reduce(
-      (acc, { name }) => ({
-        ...acc,
-        [name]: function () {
-          return name !== 'init'
-            ? instance.call(name, Object.values(arguments))
-            : instance.deploy(Object.values(arguments))
-        }
-      }),
-      {}
-    )
-
-  return instance
-}
-
-// @TODO Remove after compiler can decode using type from ACI
-function transformReturnType (returns, { bindings } = {}) {
-  if (typeof returns === 'string' && returns.indexOf(bindings.contractName) !== -1) {
-    returns = linkTypeDefs(returns, bindings)
-  }
-  try {
-    if (typeof returns === 'string') return returns
-    if (typeof returns === 'object') {
-      const [[key, value]] = Object.entries(returns)
-      return `${key !== 'tuple' && key !== 'record' ? key : ''}(${value
-        .reduce(
-          (acc, el, i) => {
-            if (i !== 0) acc += ','
-            acc += transformReturnType(key !== 'record' ? el : el.type, { bindings })
-            return acc
-          },
-          '')})`
-    }
-  } catch (e) {
-    return null
-  }
-}
-
-function call (self) {
-  return async function (fn, params = [], options = {}) {
-    const opt = R.merge(this.options, options)
-    const fnACI = getFunctionACI(this.aci, fn)
-    const source = opt.source || this.source
-    if (!fn) throw new Error('Function name is required')
-    if (!this.deployInfo.address) throw new Error('You need to deploy contract before calling!')
-
-    params = !opt.skipArgsConvert ? await prepareArgsForEncode(fnACI, params, { compilerVersion: this.compilerVersion }) : params
-    const result = opt.callStatic
-      ? await self.contractCallStatic(source, this.deployInfo.address, fn, params, {
-        top: opt.top,
-        opt
-      })
-      : await self.contractCall(source, this.deployInfo.address, fn, params, opt)
-    return {
-      ...result,
-      decodedResult: await transformDecodedData(
-        fnACI.returns,
-        await result.decode(),
-        { ...opt, compilerVersion: this.compilerVersion, bindings: fnACI.bindings }
-      )
-    }
-  }
-}
-
-function deploy (self) {
-  return async function (init = [], options = {}) {
-    const opt = R.merge(this.options, options)
-    const fnACI = getFunctionACI(this.aci, 'init')
-    if (!this.compiled) await this.compile()
-    init = !opt.skipArgsConvert ? await prepareArgsForEncode(fnACI, init, { compilerVersion: this.compilerVersion }) : init
-
-    const { owner, transaction, address, createdAt, result, rawTx } = await self.contractDeploy(this.compiled, opt.source || this.source, init, opt)
-    this.deployInfo = { owner, transaction, address, createdAt, result, rawTx }
-    return this
-  }
-}
-
-function compile (self) {
-  return async function () {
-    const { bytecode } = await self.contractCompile(this.source)
-    this.compiled = bytecode
-    return this
-  }
-}
-
-/**
- * Contract ACI Stamp
- *
- * @function
- * @alias module:@aeternity/aepp-sdk/es/contract/aci
- * @rtype Stamp
- * @return {Object} Contract compiler instance
- * @example ContractACI()
- */
-const ContractACI = AsyncInit.compose({
-  methods: {
-    getContractInstance
-  }
-})
-
-export default ContractACI
