+++ examples/browser/vuejs/connect-two-ae/identity/src/components/Home.vue
@@ -42,15 +42,14 @@
 
 <script>
   // AE_SDK_MODULES is a webpack alias present in webpack.config.js
-  import Wallet from '@aeternity/aepp-sdk/es/ae/wallet'
-  import MemoryAccount from '@aeternity/aepp-sdk/es/account/memory'
+  import { Wallet, MemoryAccount } from '@aeternity/aepp-sdk/es'
 
   export default {
     data () {
       return {
         runningInFrame: window.parent !== window,
-        pub: 'ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU', // Your public key
-        priv: 'a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61', // Your private key
+        pub: 'PROVIDE_YOUR_PUB', // Your public key
+        priv: 'PROVIDE_YOUR_PPRIV', // Your private key
         client: null,
         balance: null,
         height: null,
