+++ test/integration/contract.js
@@ -18,6 +18,7 @@
 import { describe, it, before } from 'mocha'
 import { configure, plan, ready } from './'
 import { decode } from '../../es/tx/builder/helpers'
+
 import * as R from 'ramda'
 
 const identityContract = `
@@ -86,7 +87,7 @@ describe('Contract', function () {
   let deployed
 
   before(async function () {
-    contract = await ready(this)
+    contract = await ready(this, true, true)
   })
 
   it('precompiled bytecode can be deployed', async () => {
